+++ DevFreela.Application/InputModels/UpdateProjectInputModel.cs
@@ -5,6 +5,6 @@
         public int Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public decimal TotalCost { get; set; }
+        public double TotalCost { get; set; }
     }
 }
